requirements.txt
pytest==7.4.1
pyyaml==5.3.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.0
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.2
httpx==0.26.0
idna==3.6
more-itertools==9.1.0
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic
pydantic-settings
pydantic_core
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-ulid==1.1.0
redis==4.6.0
redis-om==0.2.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
types-pyOpenSSL==23.3.0.20240106
types-redis==4.6.0.20240106
typing_extensions==4.9.0
uvicorn==0.26.0
